% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grid_derivOnly.R
\name{build_grid_derivOnly}
\alias{build_grid_derivOnly}
\title{Build grid of values for the tempering scalar along with the
(first and second) derivatives of the log-normalising constant.}
\usage{
build_grid_derivOnly(
  compiled.model.prior,
  eps = 0.01,
  M = 10,
  J = 15,
  v1 = 5,
  v2 = 10,
  stan.list,
  pars,
  niter = 5000,
  strict = FALSE,
  step_size = 0.99,
  tdepth = 15
)
}
\arguments{
\item{compiled.model.prior}{a \code{stan_model} object containing the tempered target.}

\item{eps}{grid increment factor. Default is \code{eps=0.01}.}

\item{M}{maximum value for the tempering scalar. Default is \code{M=1}.}

\item{J}{number of grid points. Default is \code{J=15}.}

\item{v1}{scalar that controls the grid spacing in the minimum-finding phase.
Default is \code{v1 = 5}.}

\item{v2}{scalar that controls the grid spacing in the "gap-plugging" phase.
Default is \code{v2 = 10}.}

\item{stan.list}{list with data for the Stan model.}

\item{pars}{which parameters to track.}

\item{niter}{number of iterations, defaults to \code{5000}.}

\item{strict}{whether to use strict computational specs.}

\item{step_size}{if \code{spec=TRUE}, sets the step_size. Default is \code{step_size = 0.99}.}

\item{tdepth}{if \code{spec=TRUE}, sets the tree_depth. Default is \code{tdepth = 15}.}
}
\value{
a list containing a data.frame with the grid points, log-normalising constant and derivatives.
If \code{pars != NA}, MCMC summaries for selected parameters are also returned.
}
\description{
Build grid of values for the tempering scalar along with the
(first and second) derivatives of the log-normalising constant.
}
\seealso{
\code{\link[npowerPrioR]{build_grid}}
}
