% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lc_df_derivOnly.R
\name{create_lc_df_derivOnly}
\alias{create_lc_df_derivOnly}
\title{Create a data.frame with the log-normalising derivatives (only) for a vector of values of the tempering factor.}
\usage{
create_lc_df_derivOnly(
  a0_grid,
  compiled.model.prior,
  stan.list,
  pars = NA,
  niter = 5000,
  strict = FALSE,
  step_size = 0.99,
  tdepth = 15
)
}
\arguments{
\item{a0_grid}{a vector of values of the tempering factor.}

\item{compiled.model.prior}{a \code{stan_model} object containing the tempered target.}

\item{stan.list}{list with data for the Stan model.}

\item{pars}{which parameters to track.}

\item{niter}{number of iterations, defaults to \code{5000}.}

\item{strict}{whether to use strict computational specs.}

\item{step_size}{if \code{spec=TRUE}, sets the step_size. Default is \code{step_size = 0.99}.}

\item{tdepth}{if \code{spec=TRUE}, sets the tree_depth. Default is \code{tdepth = 15}.}
}
\value{
a list containing a data.frame with the grid points, log-normalising constant and derivatives.
If \code{pars != NA}, MCMC summaries for selected parameters are also returned.
}
\description{
Create a data.frame with the log-normalising derivatives (only) for a vector of values of the tempering factor.
}
