% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_band.R
\name{get_band}
\alias{get_band}
\title{Get confidence bands from (GAM, lm) fit}
\usage{
get_band(fit, xpred, alpha = 0.95)
}
\arguments{
\item{fit}{an object containing a \code{lm} or \code{gam}, with a \code{predict} method.}

\item{xpred}{a vector with values at which to predict.}

\item{alpha}{the level of confidence to be used in a normal approximation (default is 95\%).}
}
\value{
a data.frame with mean, lower and upper predictions.
}
\description{
Get confidence bands from (GAM, lm) fit
}
