% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deriv_only.R
\name{get_deriv_only}
\alias{get_deriv_only}
\title{Compute the derivative of the log-normalising constant  for a given tempering scalar.}
\usage{
get_deriv_only(
  a,
  compiled.model.prior,
  stan.list,
  pars = NA,
  niter = 5000,
  strict = FALSE,
  step_size = 0.99,
  tdepth = 15
)
}
\arguments{
\item{a}{the tempering scalar.}

\item{compiled.model.prior}{a \code{stan_model} object containing the tempered target.}

\item{stan.list}{list with data for the Stan model.}

\item{pars}{which parameters to track.}

\item{niter}{number of iterations, defaults to \code{5000}.}

\item{strict}{whether to use strict computational specs.}

\item{step_size}{if \code{spec=TRUE}, sets the step_size. Default is \code{step_size = 0.99}.}

\item{tdepth}{if \code{spec=TRUE}, sets the tree_depth. Default is \code{tdepth = 15}.}
}
\value{
a list with the \code{stanfit} object, the first and second derivatives of the log-normalising constant.
}
\description{
Compute the derivative of the log-normalising constant  for a given tempering scalar.
}
\seealso{
\code{\link[npowerPrioR]{get_deriv_and_mal}}
}
