% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_bci}
\alias{mean_bci}
\title{Mean and Bayesian credibility interval (BCI)}
\usage{
mean_bci(x, alpha = 0.95)
}
\arguments{
\item{x}{a vector of MCMC samples.}

\item{alpha}{a credibility level. Default is 95\%.}
}
\value{
a data frame containing the mean and alpha*100\% BCI for x.
}
\description{
Mean and Bayesian credibility interval (BCI)
}
