% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plug_gap}
\alias{plug_gap}
\title{Plug the gaps in a vector of a0s.}
\usage{
plug_gap(x)
}
\arguments{
\item{x}{a vector of a0s.}
}
\value{
a new vector a0s with "gaps" filled in.
}
\description{
Plug the gaps in a vector of a0s.
}
