% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarise_posterior}
\alias{summarise_posterior}
\title{Obtain all summaries for stanfit}
\usage{
summarise_posterior(obj)
}
\arguments{
\item{obj}{a stanfit object.}
}
\value{
a data.frame with mean and BCI for each 'parameter' in \code{obj}
}
\description{
Obtain all summaries for stanfit
}
